requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.5
django-phonenumber-field==4.0.0
phonenumbers==8.12.2
Pillow==7.1.2
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1



